Adaptions to allow -Werror compilation (warnings as errors)
- mostly by making type casts explicit
- using `__strftimelike()` macro (but unfortunately, support at least up to gcc 11.2 is still broken and we need `-Wno-format-nonliteral` to get through with `-Werror`)

+++ utils.cpp
@@ -653,10 +653,10 @@ int p44::gtinCheckDigit(uint64_t aGtin)
   // - sum of digits*3 at odd digit positions (least significant=rightmost=1=odd) + sum of digits at even positions.
   // - check digit is the value to add to sum to get an even multiple of 10
   int sum = 0;
-  int oldcheck = aGtin % 10; // current check digit as found in aGtin
+  int oldcheck = (int)(aGtin % 10); // current check digit as found in aGtin
   for (int i=0; i<20; i++) {
     aGtin /= 10;
-    int dig = aGtin % 10;
+    int dig = (int)(aGtin % 10);
     sum += ((i&1)==0 ? 3 : 1) * dig;
   }
   int newcheck = sum%10;
